requirements.txt
# Streamlit app
streamlit==1.35.0

# Audio recording support
streamlit-audiorecorder>=0.1.2

# OpenAI API
openai>=1.0.0

# Data handling
pandas>=1.3.0
numpy>=1.21.0

# ML via PyCaret
pycaret[full]==3.3.0
scikit-learn>=1.0.2
tqdm>=4.64.0

# Extra utilities
requests>=2.25.1
